+++ src/localization/index.ts
@@ -1,1 +1,2 @@
 export * from './localization.enum';
+export * from './localize';