+++ src/panelWebView/components/Fields/WrapperField.tsx
@@ -216,6 +216,7 @@ export const WrapperField: React.FunctionComponent<IWrapperFieldProps> = ({
           value={(fieldValue as string) || null}
           required={!!field.required}
           settings={settings}
+          action={field.action}
         />
       </FieldBoundary>
     );
@@ -307,6 +308,7 @@ export const WrapperField: React.FunctionComponent<IWrapperFieldProps> = ({
           limit={field.taxonomyLimit}
           renderAsString={field.singleValueAsString}
           required={!!field.required}
+          action={field.action}
         />
       </FieldBoundary>
     );