+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ New features
 
 - [#823](https://github.com/estruyf/vscode-front-matter/issues/823): Integrated GitHub Copilot support for titles, descriptions, and tags
+- [#824](https://github.com/estruyf/vscode-front-matter/issues/824): Added the ability to link custom actions to fields
 
 ### 🎨 Enhancements